# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Ari Water Commerce & Operations Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  admin: 'admin',
  customer: 'customer',
} as const;

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: AuthUserRole;
}

export interface Product {
  id: number;
  name: string;
  sku: string;
  /** @nullable */
  description?: string | null;
  packSize: string;
  priceKes: number;
  stockQuantity: number;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  category: string;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  sku: string;
  /** @nullable */
  description?: string | null;
  packSize: string;
  priceKes: number;
  stockQuantity: number;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
  category: string;
}

export interface ProductUpdate {
  name?: string;
  sku?: string;
  /** @nullable */
  description?: string | null;
  packSize?: string;
  priceKes?: number;
  stockQuantity?: number;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
  category?: string;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  packSize: string;
  quantity: number;
  unitPriceKes: number;
}

export interface OrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  received: 'received',
  processing: 'processing',
  dispatched: 'dispatched',
  delivered: 'delivered',
} as const;

export type OrderPaymentStatus = typeof OrderPaymentStatus[keyof typeof OrderPaymentStatus];


export const OrderPaymentStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

/**
 * @nullable
 */
export type OrderPaymentMethod = typeof OrderPaymentMethod[keyof typeof OrderPaymentMethod] | null;


export const OrderPaymentMethod = {
  mpesa: 'mpesa',
  card: 'card',
  bank_transfer: 'bank_transfer',
  null: 'null',
} as const;

export interface Order {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  status: OrderStatus;
  totalKes: number;
  deliveryAddress: string;
  phone: string;
  /** @nullable */
  notes?: string | null;
  paymentStatus: OrderPaymentStatus;
  /** @nullable */
  paymentMethod?: OrderPaymentMethod;
  /** @nullable */
  paystackRef?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type OrderDetailStatus = typeof OrderDetailStatus[keyof typeof OrderDetailStatus];


export const OrderDetailStatus = {
  received: 'received',
  processing: 'processing',
  dispatched: 'dispatched',
  delivered: 'delivered',
} as const;

export type OrderDetailPaymentStatus = typeof OrderDetailPaymentStatus[keyof typeof OrderDetailPaymentStatus];


export const OrderDetailPaymentStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Review {
  id: number;
  orderId: number;
  customerId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  remark?: string | null;
  createdAt: string;
}

export interface OrderDetail {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  status: OrderDetailStatus;
  totalKes: number;
  deliveryAddress: string;
  phone: string;
  /** @nullable */
  notes?: string | null;
  paymentStatus: OrderDetailPaymentStatus;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  paystackRef?: string | null;
  createdAt: string;
  updatedAt: string;
  items: OrderItem[];
  review?: Review;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
}

export type OrderInputPaymentMethod = typeof OrderInputPaymentMethod[keyof typeof OrderInputPaymentMethod];


export const OrderInputPaymentMethod = {
  mpesa: 'mpesa',
  card: 'card',
  bank_transfer: 'bank_transfer',
} as const;

export interface OrderInput {
  deliveryAddress: string;
  phone: string;
  /** @nullable */
  notes?: string | null;
  paymentMethod?: OrderInputPaymentMethod;
  /** @minItems 1 */
  items: OrderItemInput[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  received: 'received',
  processing: 'processing',
  dispatched: 'dispatched',
  delivered: 'delivered',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface ReviewInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  remark?: string | null;
}

export type DashboardSummaryOrdersByStatus = {
  received: number;
  processing: number;
  dispatched: number;
  delivered: number;
};

export interface DashboardSummary {
  todayRevenue: number;
  todayOrders: number;
  weekRevenue: number;
  weekOrders: number;
  totalRevenue: number;
  totalOrders: number;
  ordersByStatus: DashboardSummaryOrdersByStatus;
  lowStockProducts: number;
}

export type RecentOrderStatus = typeof RecentOrderStatus[keyof typeof RecentOrderStatus];


export const RecentOrderStatus = {
  received: 'received',
  processing: 'processing',
  dispatched: 'dispatched',
  delivered: 'delivered',
} as const;

export type RecentOrderPaymentStatus = typeof RecentOrderPaymentStatus[keyof typeof RecentOrderPaymentStatus];


export const RecentOrderPaymentStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface RecentOrder {
  id: number;
  customerName: string;
  /** @nullable */
  customerEmail?: string | null;
  totalKes: number;
  status: RecentOrderStatus;
  paymentStatus: RecentOrderPaymentStatus;
  createdAt: string;
  itemCount: number;
}

export interface RevenuePoint {
  date: string;
  revenue: number;
  orders: number;
}

export interface PaymentInitInput {
  orderId: number;
}

export interface PaymentInitResponse {
  authorizationUrl: string;
  reference: string;
}

export interface PaymentVerifyInput {
  reference: string;
}

export interface PaymentVerifyResponse {
  success: boolean;
  status: string;
  /** @nullable */
  orderId?: number | null;
}

export type ListProductsParams = {
category?: string;
inStock?: ListProductsInStock;
};

export type ListProductsInStock = typeof ListProductsInStock[keyof typeof ListProductsInStock];


export const ListProductsInStock = {
  true: 'true',
  false: 'false',
} as const;

export type ListOrdersParams = {
status?: ListOrdersStatus;
page?: number;
limit?: number;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  received: 'received',
  processing: 'processing',
  dispatched: 'dispatched',
  delivered: 'delivered',
} as const;