# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ari Water Commerce & Operations Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in and receive a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Products ──────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin sees all; customer sees own)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [received, processing, dispatched, delivered]
          required: false
        - in: query
          name: page
          schema:
            type: integer
          required: false
        - in: query
          name: limit
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order with line items
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/review:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a delivered order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ── Dashboard (admin) ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # Auth
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        phone:
          type: ["string", "null"]
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, customer]

    # Products
...
    OrderItem:
      type: object
      required: [id, productId, productName, packSize, quantity, unitPriceKes]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        packSize:
          type: string
        quantity:
          type: integer
        unitPriceKes:
          type: number
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
    Order:
      type: object
      required: [id, customerId, status, totalKes, deliveryAddress, phone, paymentStatus, createdAt, updatedAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [received, processing, dispatched, delivered]
        totalKes:
          type: number
        deliveryAddress:
          type: string
        phone:
          type: string
        notes:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [pending, completed, failed]
        paymentMethod:
          type: ["string", "null"]
          enum: [mpesa, card, bank_transfer, "null"]
        paystackRef:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderDetail:
      type: object
      required: [id, customerId, status, totalKes, deliveryAddress, phone, paymentStatus, createdAt, updatedAt, items]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        status:
          type: string
          enum: [received, processing, dispatched, delivered]
        totalKes:
          type: number
        deliveryAddress:
          type: string
        phone:
          type: string
        notes:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [pending, completed, failed]
        paymentMethod:
          type: ["string", "null"]
        paystackRef:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        review:
          $ref: "#/components/schemas/Review"
    OrderListResponse:
      type: object
      required: [orders, total, page, limit]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    OrderInput:
      type: object
      required: [deliveryAddress, phone, items]
      properties:
        deliveryAddress:
          type: string
        phone:
          type: string
        notes:
          type: ["string", "null"]
        paymentMethod:
          type: string
          enum: [mpesa, card, bank_transfer]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
          minItems: 1
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [received, processing, dispatched, delivered]

    # Reviews
    Review:
      type: object
      required: [id, orderId, customerId, rating, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        customerId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        remark:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        remark:
          type: ["string", "null"]

    # Dashboard
...