# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Ari Water Commerce & Operations Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from './authUserRole';

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: AuthUserRole;
}