# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Ari Water Commerce & Operations Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderPaymentStatus } from './orderPaymentStatus';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  status: OrderStatus;
  totalKes: number;
  deliveryAddress: string;
  phone: string;
  /** @nullable */
  notes?: string | null;
  paymentStatus: OrderPaymentStatus;
  /** @nullable */
  paymentMethod?: OrderPaymentMethod;
  /** @nullable */
  paystackRef?: string | null;
  createdAt: Date;
  updatedAt: Date;
}