# lib/api-zod/src/generated/types/orderDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Ari Water Commerce & Operations Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderDetailPaymentStatus } from './orderDetailPaymentStatus';
import type { OrderDetailStatus } from './orderDetailStatus';
import type { OrderItem } from './orderItem';
import type { Review } from './review';

export interface OrderDetail {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  status: OrderDetailStatus;
  totalKes: number;
  deliveryAddress: string;
  phone: string;
  /** @nullable */
  notes?: string | null;
  paymentStatus: OrderDetailPaymentStatus;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  paystackRef?: string | null;
  createdAt: Date;
  updatedAt: Date;
  items: OrderItem[];
  review?: Review;
}