# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Ari Water Commerce & Operations Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  packSize: string;
  quantity: number;
  unitPriceKes: number;
}