# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Ari Water Commerce & Operations Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  orderId: number;
  customerId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  remark?: string | null;
  createdAt: Date;
}